create csvparser and edit methods and variables accordingly
