+++ dino_catalog/dinodex.rb
@@ -1,55 +1,11 @@
-require 'csv'
-require 'json'
 require_relative 'dinosaur.rb'
-
-CSV::Converters[:blank_to_nil] = lambda do |field|
-  field && field.empty? ? nil : field
-end
-
-CSV_HEADER_CONVERTERS = {
-                          headers: true,
-                          header_converters: :symbol,
-                          converters: [:all, :blank_to_nil],
-                        }
+require_relative 'csvparser.rb'
 
 class Dinodex
   attr_reader :master_dinosaur_list
 
   def initialize(master_dinosaur_list)
-    @master_dinosaur_list = master_dinosaur_list || []
-  end
-
-  def african_dino_import
-    CSV.foreach("african_dinosaur_export.csv", CSV_HEADER_CONVERTERS) do |dino|
-      african_dino = {
-        name: dino[:genus],
-        period: dino[:period],
-        continent: "Africa",
-        diet: carnivore_to_diet(dino[:carnivore]),
-        weight_in_lbs: dino[:weight],
-        walking: dino[:walking],
-      }
-      @master_dinosaur_list << Dinosaur.new(african_dino)
-    end
-  end
-
-  def carnivore_to_diet(diet)
-    diet == "Yes" ? "Carnivore" : "Herbivore"
-  end
-
-  def dino_import
-    CSV.foreach("dinodex.csv", CSV_HEADER_CONVERTERS) do |dino|
-      dinosaur = {
-        name: dino[:name],
-        period: dino[:period],
-        continent: dino[:continent],
-        diet: dino[:diet],
-        weight_in_lbs: dino[:weight_in_lbs],
-        walking: dino[:walking],
-        description: dino[:description],
-      }
-      @master_dinosaur_list << Dinosaur.new(dinosaur)
-    end
+    @master_dinosaur_list = CSVParser.new.generate_master_dinosaur_list || []
   end
 
   def biped_search
